Use alternative histogram creator
